third_party/WebKit/LayoutTests/editing/selection/DOMSelection-DocumentType-expected.txt

Test to check if setBaseAndExtent guard node with null owner document (Bug 31680)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS sel.setBaseAndExtent(docType, 0, null, 0) threw exception InvalidNodeTypeError: Failed to execute 'setBaseAndExtent' on 'Selection': The node provided is of type 'c'..
PASS sel.anchorNode is null
PASS sel.anchorNode is null
PASS sel.collapse(docType) threw exception InvalidNodeTypeError: Failed to execute 'collapse' on 'Selection': The node provided is of type 'c'..
PASS sel.anchorNode is null
PASS sel.selectAllChildren(docType) threw exception InvalidNodeTypeError: Failed to execute 'selectAllChildren' on 'Selection': The node provided is of type 'c'..
PASS sel.anchorNode is null
PASS sel.anchorNode is null
PASS sel.anchorNode is null
PASS sel.containsNode(docType) is false
PASS successfullyParsed is true

TEST COMPLETE
